Fix a deadlock-causing race-condition in `TagAlongThread`:
The race - "_T_" is the `TagAlongThread`, "_C_" is the caller thread:
- _T_ waits on `_iteration_trigger`.
- _C_ sets `_iteration_trigger`.
- _T_ sets `_iterating`.
- _T_ performs the loop too fast - before _C_ has a chance to do anything.
- _T_ clears `_iterating`.
- _C_ waits on `_iterating`.
- _T_ starts over again - and waits on `_iteration_trigger`.

_C_ is waiting for _T_ to set `_iterating` - but _T_ has already set it
and cleared it, and won't set it again until someone else sets
`_iteration_trigger`. This may never happen because _C_ is stuck and the
program can't progress...

The solution that every second _C_ will look at `_not_iterating` - an
`Event` that we already have for a different purpose and gets set after
the iteration is done.

Now, in theory, a deadlock is still possible - but it would require
exact timing that must repeat every second, so we should be safe.